momentum/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
ccxt>=4.3.30
pyyaml>=6.0.1
